requirements.txt
# Core ML libraries
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# Stable Baselines3 for reference implementations
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Visualization and plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# UI and interactive tools
streamlit>=1.25.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
tqdm>=4.65.0
rich>=13.0.0
